backend/requirements.txt
# AI Air Quality Agent Backend Requirements

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data Processing
pandas==2.0.3
numpy==1.24.3

# AI/ML
google-generativeai

# Environment Variables
python-dotenv==1.0.0

# Logging and Utilities
python-dateutil==2.8.2

# Optional: Machine Learning (uncomment if needed)
# scikit-learn==1.3.0
# tensorflow==2.13.0
# torch==2.0.1

# Optional: Database (uncomment if needed)
# sqlalchemy==2.0.19
# psycopg2-binary==2.9.7

# Optional: API clients (uncomment if needed)
# requests==2.31.0
# aiohttp==3.8.5

# Development and Testing
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0
